u-u: implement plugin system with `postrun()` functionality
This commit implements a new plugin system. Plugins are python
classes that start with `UnattendedUpgradesPlugin` and that are
locates in one of the following directories:
```
/etc/unattended-upgrades/plugins/
/usr/share/unattended-upgrades/plugins
```
Plugins are also searched in the `UNATTENDED_UPGRADES_PLUGIN_PATH`
environment path and in any of the directories in the apt configration:
`Unattended-Upgrade::Dirs::Plugins::`.

The `postrun()` function is called with a single python dictionary
as the argument. This dictionary contains the following elements:
* "plugin-api": the API version as string of the form "1.0"
* "hostname": The hostname of the machine that run u-u.
* "success": A boolean that indicates if the run was successful
* "result-str": A string with a human readable (and translated) status message
* "packages-upgraded": A list of packages that got upgraded.
* "packages-kept-back": A list of packages kept back.
* "packages-kept-installed": A list of packages not auto-removed.
* "reboot-required": Indicates a reboot is required.
* "log-dpkg": The full dpkg log output.
* "log-unattended-upgrades": The full unattended-upgrades log.

This new code can be tested with:
```
$ sudo PYTHONPATH=. UNATTENDED_UPGRADES_PLUGIN_PATH=./examples/plugins/ ./unattended-upgrade
```